Move files into src/ and update references
Move static assets, CSS and JS under a new src/ directory and update all references accordingly. HTML pages now load resources from ./src/; unit tests, README/docs, ESLint and Vitest coverage globs updated to use src/js paths; and Vite base set to "./" to support relative builds. This reorganizes the project structure and updates tooling and tests to match the new layout.
