+++ about.html
@@ -4,15 +4,15 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>About - Pomodoro Web</title>
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="icon" href="assets/icon/icon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="./src/css/style.css">
+    <link rel="icon" href="./src/assets/icon/icon.ico" type="image/x-icon">
 </head>
 <body>
     <header class="header">
         <div class="header-container">
             <div class="logo-container">
                 <a href="index.html" class="logo-link">
-                    <img src="assets/icon/icon.ico" alt="Pomodoro Logo" class="logo-img">
+                    <img src="./src/assets/icon/icon.ico" alt="Pomodoro Logo" class="logo-img">
                 </a>
                 <h1 class="logo">Pomodoro Timer</h1>
             </div>
@@ -67,7 +67,7 @@
             </p>
 
             <h3>About Me</h3>
-            <img src="assets/pictures/author.jpg" alt="Author" class="about-photo">
+            <img src="./src/assets/pictures/author.jpg" alt="Author" class="about-photo">
             <p>
                 I'm Oleh, a developer who created this app to help myself and others who have trouble staying focused on tasks.
                 I'm a student at the Lviv Polytechnic National University, and I enjoy building tools that make life easier.
@@ -90,6 +90,6 @@
         </div>
     </footer>
 
-    <script type="module" src="js/entries/secondary-page.js"></script>
+    <script type="module" src="./src/js/entries/secondary-page.js"></script>
 </body>
 </html>
