+++ helpus.html
@@ -4,15 +4,15 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Help Us - Pomodoro Web</title>
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="icon" href="assets/icon/icon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="./src/css/style.css">
+    <link rel="icon" href="./src/assets/icon/icon.ico" type="image/x-icon">
 </head>
 <body>
     <header class="header">
         <div class="header-container">
             <div class="logo-container">
                 <a href="index.html" class="logo-link">
-                    <img src="assets/icon/icon.ico" alt="Pomodoro Logo" class="logo-img">
+                    <img src="./src/assets/icon/icon.ico" alt="Pomodoro Logo" class="logo-img">
                 </a>
                 <h1 class="logo">Pomodoro Timer</h1>
             </div>
@@ -49,7 +49,7 @@
                 Scan the QR code below or <a href="https://send.monobank.ua/jar/8NC2t9efmF" target="_blank" rel="noopener noreferrer">follow this link</a>.
             </p>
             <div class="qr-code-container">
-                <img src="assets/pictures/monobank.jpeg" alt="QR Code for Monobank Donation" class="qr-code">
+                <img src="./src/assets/pictures/monobank.jpeg" alt="QR Code for Monobank Donation" class="qr-code">
             </div>
 
             <p>
@@ -70,6 +70,6 @@
         </div>
     </footer>
 
-    <script type="module" src="js/entries/secondary-page.js"></script>
+    <script type="module" src="./src/js/entries/secondary-page.js"></script>
 </body>
 </html>
