+++ src/css/modules/timer.css
