+++ src/js/entries/timer-page.js
