+++ src/js/mini-timer.js
