+++ src/js/timer-state.js
