+++ src/js/timer.js
